requirements.txt
streamlit
pandas
numpy
python-docx
PyPDF2
Pillow
pytesseract
python-dotenv
plotly
scipy
scikit-learn
statsmodels
seaborn
google-generativeai
opencv-python
matplotlib